+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 2.0-SNAPSHOT
 
 * #12 Update to Kotlin 1.4
+* #13 Config for prefix 'xyz' not found.
 
 ## 1.0.RC1